move main renderer contexts to contexts/main
