+++ src/renderer/screens/Home/Events/Item.tsx
@@ -13,7 +13,7 @@ import { getAddressNonce } from '@/utils/AccountUtils';
 import { isValidHttpUrl } from '@w3ux/utils';
 import { Identicon } from '@app/library/Identicon';
 import { useEffect, useState, memo } from 'react';
-import { useEvents } from '@/renderer/contexts/Events';
+import { useEvents } from '@/renderer/contexts/main/Events';
 import { useBootstrapping } from '@/renderer/contexts/Bootstrapping';
 import { useTooltip } from '@/renderer/contexts/common/Tooltip';
 import type { EventAccountData } from '@/types/reporter';
