+++ src/renderer/screens/Home/Events/index.tsx
@@ -1,7 +1,7 @@
 // Copyright 2024 @rossbulat/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { useEvents } from '@app/contexts/Events';
+import { useEvents } from '@/renderer/contexts/main/Events';
 import { useState, useMemo } from 'react';
 import { Category } from './Category';
 import { NoEvents } from './NoEvents';
